v0.1.30
1.30 - re write of UI

Altered creation of PEQ filters so multiple shelf and pass filters as well as well as peak filters can be created
Changed UI with respect to creation and deletion of filters so that the precise filter to be deleted can be selected
Added an apply button which creates a manual save of settings rather than doing automatically - simplifies code and follows UX standards
Added a load button for presets - simplifies code and follows UX standards
Changed design so that creation of filters is entirely on the Web page with perl script now just a persistence layer - simplifies code and reduces flashing messages, also makes REW load more reliable.
Added a feature so that wav impulses can be imported and are saved to the impulse folder using same mechanism as REW files.
fixed the pre-amp text field, so it is now editable
added logic to pause and restart the currently playing song 0.5s ahead of current time, when settings are saved. This forces reload of settings by the DSP engine, so that settings are changed in real-time. This feature does not work on radio streams (because it may break the stream) and wma files as pausing and restarting does not work with them.
split the perl and javascript libraries into functionally related code for easier maintenance.
applied a significant restyle to the page, in order to keep a modern clean interface
amended binary to cope with new PEQ structure.
NB the code is backwards compatible with the old settings and these are converted on load to the UI.